Adding a new script (it needs a new name though)
PSstart is a script to start a process with a specified core affinity and scheduler priority.  This can be helpful with legacy applications which make poor use of multithreading and will schedule threads across CPUs needlessly for no performance gain.

It can be called from cmd or from a shortcut, which is ideal for launching programs with that issue.